Revert f0d367dd4a to avoid ORA-03127 on ANALYZE
Commit f0d367dd4a introduced code to stop reading a LOB if we got enough
data for ANALYZE.  That commit was a mitigation for a performance regression
introduced in 7d546162b1, which told OCILobRead2 to fetch the whole LOB to
avoid a crash.

However, it seems that Oracle insists that you finish reading the LOB to the
end once you have started reading it in polling mode.  Otherwise, you get

ORA-03127: no new operations allowed until the active operation ends

So we have to revert f0d367dd4a and accept the performance regression during
ANALYZE.

This closes #758 reported by "MMFERNANDEZC".